% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\docType{class}
\name{Builder}
\alias{Builder}
\title{Class providing methods for building connection object}
\format{\code{\link{R6Class}} object.}
\usage{
Builder
}
\value{
Object of \code{\link{R6Class}} with methods for communication with a database (server)
}
\description{
Class providing methods for building connection object
}
\section{Fields}{

\describe{
\item{\code{serveraddress}}{Stores address of your lightning server.}

\item{\code{sessionid}}{Stores id of your current session on the server.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/ChrisNice89/ORM}
  \item{\code{new(provider=c(MySql,TsQl,MsAccess), path,dbpassword="")}}{This method creates a builder \code{instance}.}
  \item{\code{addCredendials(username="",password="")}}{This method is used to create object of user specified credentials}
  \item{\code{build()}}{Creates specified connection to the database (server)}}
}

\examples{
b<-Builder$new(provider = "MySql")
cnn<-b$build()

b<-Builder$new(provider = "MySql")
cnn<-b$addCredentials(username = "Admin",password = "SesameOpen")$build()

}
\seealso{
Other sql: \code{\link{Credentials}},
  \code{\link{factory}}, \code{\link{sqlConnection}}
}
\concept{sql}
\keyword{data}
