% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFactory.R
\docType{class}
\name{connectionFactory}
\alias{connectionFactory}
\title{SqlFactory liefert mehoden zum erstellen eines builders (Entwurfsmuster)}
\format{\code{\link{R6Class}} object.}
\usage{
connectionFactory()
}
\value{
Object of \code{\link{R6Class}} xxxxxxxxxxxxxxxxxxxx (x)
}
\description{
SqlFactory liefert mehoden zum erstellen eines builders (Entwurfsmuster)
}
\section{Fields}{

\describe{
\item{\code{x}}{blabla.}

\item{\code{y}}{blabla.}
}}

\section{Construction}{

```
xxxxxxxxxxxxxxxxxxxxxx
```
}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/ChrisNice89/rdao}
  \item{\code{new()}}{xxx \code{Factory}.}
  \item{\code{xx(yy,yy="")}}{xx \code{zz}.}
  \item{\code{zz(yy,yy="")}}{xx \code{zz}.}}
}

\examples{
xxxxxxxxxxxxxxxxxxxxxx


}
\seealso{
Other sql: \code{\link{Builder}},
  \code{\link{Credentials}}, \code{\link{IDiamonds}},
  \code{\link{Validator}}, \code{\link{businessLogic}},
  \code{\link{diamondsFactory}}, \code{\link{diamond}},
  \code{\link{sqlCommand}}, \code{\link{sqlConnection}},
  \code{\link{sqlResult}}
}
\concept{sql}
\keyword{data}
