% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\docType{class}
\name{Credentials}
\alias{Credentials}
\title{Class providing methods for creating a credential instance}
\format{\code{\link{R6Class}} object.}
\usage{
Credentials
}
\value{
Object of \code{\link{R6Class}} with methods user specified credentials (server)
}
\description{
Class providing methods for creating a credential instance
}
\section{Fields}{

\describe{
\item{\code{username}}{Readonly: Returns user name.}

\item{\code{password}}{Readonly: Returns user password.}
}}

\section{Methods}{

\describe{}
}

\examples{
c<-Credentials$new(username = "Admin",password = "SesameOpen")

}
\seealso{
Other sql: \code{\link{Builder}}, \code{\link{factory}},
  \code{\link{sqlCommand}}, \code{\link{sqlConnection}}
}
\concept{sql}
\keyword{data}
