% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlFactory.R
\docType{class}
\name{factory}
\alias{factory}
\title{Class providing methods for creating a builder instance
Factory methodes are the access to the implementation (Builder class)}
\format{\code{\link{R6Class}} object.}
\usage{
factory()
}
\value{
Object of \code{\link{R6Class}} with methods for communication with a database (server)
}
\description{
Class providing methods for creating a builder instance
Factory methodes are the access to the implementation (Builder class)
}
\section{Fields}{

\describe{
\item{\code{serveraddress}}{Stores address of your lightning server.}

\item{\code{sessionid}}{Stores id of your current session on the server.}
}}

\section{Construction}{

```
f<-factory()
```
}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/ChrisNice89/ORM}
  \item{\code{new()}}{This method creates a instance \code{Factory}.}
  \item{\code{mySql(path,dbpassword="")}}{This method is used to create a builder instance for \code{sqlConnection}.}
  \item{\code{msAccess(path,dbpassword="")}}{This method is used to create a builder instance for \code{sqlConnection}.}}
}

\examples{
f<-sqlFactory$new()
b<-f$mySql(path,dbpassword="")

}
\seealso{
Other sql: \code{\link{Builder}},
  \code{\link{Credentials}}, \code{\link{sqlConnection}}
}
\concept{sql}
\keyword{data}
