% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_rdao.R
\docType{package}
\name{rdao}
\alias{rdao}
\alias{rdao-package}
\title{rdao}
\description{
rdao
}
\section{Classes}{


\describe{
  \item{SqlFactory}{A database framework. (\link{SqlFactory})}
  \item{SqlConnection}{A wrapper around connection implementations. (\link{SqlConnection})}
}
}

