% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_rorm.R
\docType{package}
\name{rorm}
\alias{rorm}
\alias{rorm-package}
\title{rorm}
\description{
rorm
}
\section{Classes}{


\describe{
  \item{SqlFactory}{A database framework. (\link{SqlFactory})}
  \item{SqlConnection}{A wrapper around connection implementations. (\link{SqlConnection})}
}
}

