% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlCommand.R
\docType{class}
\name{sqlCommand}
\alias{sqlCommand}
\title{Class providing some methods to for CRUD actions}
\format{\code{\link{R6Class}} object.}
\usage{
sqlCommand
}
\value{
Object of \code{\link{R6Class}} with methods for communication with a database (server)
}
\description{
Class providing some methods to for CRUD actions
}
\section{Fields}{

\describe{
\item{\code{}}{}
}}

\section{Construction}{

```
sqlConnection$new()
```
}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/ChrisNice89/ORM}
  \item{\code{new()}}{This method creates a sqlConnection \code{instance}.}
  \item{\code{isString((x, objnm = deparse(substitute(x))))}}{Check if x is numeric}
  \item{\code{isNumeric((x, objnm = deparse(substitute(x))))}}{Check if x is numeric}}
}

\examples{
# cnn<-sqlConnection$new(connectionstring = "myconnectionstring")
}
\seealso{
Other sql: \code{\link{Builder}},
  \code{\link{Credentials}}, \code{\link{factory}},
  \code{\link{sqlConnection}}
}
\concept{sql}
\keyword{command}
