% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlInterface.R
\docType{class}
\name{sqlInterface}
\alias{sqlInterface}
\title{Abstrakte Klasse implementiert Interface}
\format{\code{\link{R6Class}} object.}
\usage{
sqlInterface
}
\value{
Object of \code{\link{R6Class}} xxxxxxxxxxxxxxxxxxxx (x)
}
\description{
Abstrakte Klasse implementiert Interface
}
\section{Fields}{

\describe{
\item{\code{x}}{blabla.}

\item{\code{y}}{blabla.}
}}

\section{Construction}{

```
xxxxxxxxxxxxxxxxxxxxxx
```
}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/ChrisNice89/rdao}
  \item{\code{new()}}{xxx \code{Factory}.}
  \item{\code{xx(yy,yy="")}}{xx \code{zz}.}
  \item{\code{zz(yy,yy="")}}{xx \code{zz}.}}
}

\examples{
xxxxxxxxxxxxxxxxxxxxxx


}
\seealso{
Other sql: \code{\link{Builder}},
  \code{\link{Credentials}}, \code{\link{IDiamonds}},
  \code{\link{Validator}}, \code{\link{businessLogic}},
  \code{\link{connectionFactory}},
  \code{\link{diamondsFactory}}, \code{\link{diamond}},
  \code{\link{sqlCommand}}, \code{\link{sqlConnection}},
  \code{\link{sqlResult}}
}
\concept{sql}
\keyword{data}
