% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{class}
\name{Validator}
\alias{Validator}
\title{Class providing some methods to validate data}
\format{\code{\link{R6Class}} object.}
\usage{
Validator
}
\value{
Object of \code{\link{R6Class}} with methods for communication with a database (server)
}
\description{
Class providing some methods to validate data
}
\section{Fields}{

\describe{
\item{\code{sessionid}}{Stores id of your current session on the server.}
}}

\section{Construction}{

```
Validator$new()
```
}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/ChrisNice89/ORM}
  \item{\code{new()}}{This method creates a validator \code{instance}.}
  \item{\code{isString((x, objnm = deparse(substitute(x))))}}{Check if x is numeric}
  \item{\code{isNumeric((x, objnm = deparse(substitute(x))))}}{Check if x is numeric}}
}

\examples{
validate<-Validator$new()
validate$isNumerix("1") -> throws error

}
\seealso{
Other utils: \code{\link{HashMap}}
}
\concept{utils}
\keyword{checks}
